requirements.txt
aiofiles
fastapi
jinja2
markdown2
openai
python-dotenv
starlette
typing-extensions
uvicorn
itsdangerous
motor
pymongo
uvicorn
boto3